fixed compiler warnings

+++ frontends/lexer.c
@@ -440,7 +440,7 @@ isCIdentifierChar(int c)
  * actual parsing functions
  */
 static int
-parseNumber(LexStream *L, unsigned int base, uint32_t *num)
+parseNumber(LexStream *L, unsigned int base, uint64_t *num)
 {
     uint32_t uval, digit;
     uint64_t tenval;
